refactor: 이달의 차트 컴포넌트 위치 변경
