+++ src/components/ChartOfTheMonth/ChartPage.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import TabMenu from "./components/TabMenu/TabMenu";
 import IdolList from "./components/IdolList/IdolList";
 import LoadMoreButton from "./components/LoadMoreButton/LoadMoreButton";
-import { FEMALE } from "../../../constants/tabGenderTypes";
+import { FEMALE } from "../../constants/tabGenderTypes";
 import { mockIdolData } from "./mockData"; // 임시 Mock 데이터 import
 import styles from "./ChartPage.module.css";
