+++ src/components/ChartOfTheMonth/components/IdolElement/IdolElement.jsx
