+++ src/components/ChartOfTheMonth/components/IdolElement/IdolElement.module.css
