+++ src/components/ChartOfTheMonth/components/IdolList/IdolList.jsx
