+++ src/components/ChartOfTheMonth/components/IdolList/IdolList.module.css
