+++ src/components/ChartOfTheMonth/components/TabMenu/TabMenu.module.css
