+++ src/pages/ListPage/ListPage.jsx
@@ -1,0 +1,7 @@
+import ChartPage from "../../components/ChartOfTheMonth/ChartPage";
+
+const ListPage = () => {
+  return <ChartPage />;
+};
+
+export default ListPage;
